requirements.txt
streamlit
pillow
pycryptodome
streamlit-lottie
requests